web/requirements.txt
cassandra-driver
futures==3.0.3
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
six==1.9.0
Werkzeug==0.10.4
Flask==1.0.2
